FIX: Moving of diagnostics to slatekit-tracking
