+++ src/lib/kotlin/slatekit-jobs/build.gradle
@@ -81,11 +81,13 @@ dependencies {
     if( slatekitSetupViaMaven ) {
         compile "com.slatekit:slatekit-results:$slatekit_version"
         compile "com.slatekit:slatekit-common:$slatekit_version"
+        compile "com.slatekit:slatekit-tracking:$slatekit_version"
         compile "com.slatekit:slatekit-functions:$slatekit_version"
     } else {
         // */
         compile project(":slatekit-result")
         compile project(":slatekit-common")
+        compile project(":slatekit-tracking")
         compile project(":slatekit-functions")
     } //</slatekit_local>
 }
