+++ src/lib/kotlin/slatekit-jobs/src/main/kotlin/slatekit/jobs/Recorder.kt
@@ -1,5 +1,5 @@
 package slatekit.jobs
 
-import slatekit.common.metrics.Recorder
+import slatekit.tracking.Recorder
 
 typealias Recorder = Recorder<Task, WorkState>
