+++ src/lib/kotlin/slatekit-providers/src/main/kotlin/slatekit/providers/metrics/DDMetrics.kt
@@ -9,7 +9,7 @@ import io.micrometer.core.instrument.binder.jvm.JvmThreadMetrics
 import io.micrometer.core.instrument.binder.system.ProcessorMetrics
 import io.micrometer.datadog.DatadogMeterRegistry
 import slatekit.common.Identity
-import slatekit.common.metrics.*
+import slatekit.tracking.*
 
 
 class DDMetrics(val registry: MeterRegistry,
