+++ src/lib/kotlin/slatekit-tracking/build.gradle
@@ -77,7 +77,7 @@ dependencies {
     compile 'com.squareup.okhttp3:okhttp:3.9.0'
     compile 'org.threeten:threetenbp:1.3.8'
 
-    /* <slatekit_local>
+    // /* <slatekit_local>
     if( slatekitSetupViaMaven ) {
         compile "com.slatekit:slatekit-results:$slatekit_version"
         compile "com.slatekit:slatekit-common:$slatekit_version"
@@ -85,7 +85,7 @@ dependencies {
         // */
         compile project(":slatekit-result")
         compile project(":slatekit-common")
-    // } //</slatekit_local>
+    } //</slatekit_local>
 }
 
 // ==================================================================
