+++ src/lib/kotlin/slatekit/build.gradle
@@ -88,6 +88,9 @@ dependencies {
         compile "com.slatekit:slatekit-common:$slatekit_version"
         compile "com.slatekit:slatekit-app:$slatekit_version"
         compile "com.slatekit:slatekit-db:$slatekit_version"
+        compile "com.slatekit:slatekit-tracking:$slatekit_version"
+        compile "com.slatekit:slatekit-functions:$slatekit_version"
+        compile "com.slatekit:slatekit-cmds:$slatekit_version"
         compile "com.slatekit:slatekit-meta:$slatekit_version"
         compile "com.slatekit:slatekit-cloud:$slatekit_version"
         compile "com.slatekit:slatekit-query:$slatekit_version"
@@ -105,13 +108,14 @@ dependencies {
         implementation project(":slatekit-common")
         implementation project(":slatekit-app")
         implementation project(":slatekit-cli")
+        implementation project(":slatekit-tracking")
+        implementation project(":slatekit-functions")
+        implementation project(":slatekit-cmds")
         implementation project(":slatekit-db")
         implementation project(":slatekit-meta")
         implementation project(":slatekit-query")
         implementation project(":slatekit-cloud")
         implementation project(":slatekit-entities")
-        implementation project(":slatekit-functions")
-        implementation project(":slatekit-cmds")
         implementation project(":slatekit-orm")
         implementation project(":slatekit-core")
         implementation project(":slatekit-notifications")
